Added CustomCurrency
