+++ lib/src/api/entities/currency/currency.dart
@@ -1,0 +1,8 @@
+import '../../../../restrr.dart';
+
+abstract class Currency extends RestrrEntity {
+  String get name;
+  String get symbol;
+  String get isoCode;
+  int get decimalPlaces;
+}
