+++ lib/src/internal/entities/currency/custom_currency_impl.dart
@@ -1,30 +1,20 @@
 import 'package:restrr/restrr.dart';
-import 'package:restrr/src/internal/entities/restrr_entity_impl.dart';
 
-class CurrencyImpl extends RestrrEntityImpl implements Currency {
-  @override
-  final String name;
-  @override
-  final String symbol;
-  @override
-  final String isoCode;
-  @override
-  final int decimalPlaces;
+import 'currency_impl.dart';
+
+class CustomCurrencyImpl extends CurrencyImpl implements CustomCurrency {
   @override
   final int? user;
 
-  const CurrencyImpl({
+  const CustomCurrencyImpl({
     required super.api,
     required super.id,
-    required this.name,
-    required this.symbol,
-    required this.isoCode,
-    required this.decimalPlaces,
+    required super.name,
+    required super.symbol,
+    required super.isoCode,
+    required super.decimalPlaces,
     required this.user,
   });
-
-  @override
-  bool get isCustom => user != null;
 
   @override
   bool isCreatedBy(User user) => this.user == user.id;
